backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
sqladmin==0.17.0
Jinja2==3.1.4
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
email-validator==2.2.0

